Remove fastlane, move build settings into .xcconfig files, update read me, turn on and fix new build warnings.
Goal here was to make project easier to use for people who aren't me: hopefully there's no longer any need to keep uncommitted changes around just to build things.

Added benefit is removing a dev dependency (fastlane) that wasn't really pulling its weight.

+++ Smilies/Framework/SmilieKeyboardView.m
@@ -323,13 +323,13 @@ static NSString * const CellIdentifier = @"SmilieCell";
 - (void)scrollViewDidEndDragging:(UIScrollView *)scrollView willDecelerate:(BOOL)decelerate
 {
     if (!decelerate) {
-        SmilieKeyboardSetScrollFractionForSmilieList(self.selectedSmilieList, scrollView.contentOffset.y / scrollView.contentSize.height);
+        SmilieKeyboardSetScrollFractionForSmilieList(self.selectedSmilieList, (float)scrollView.contentOffset.y / (float)scrollView.contentSize.height);
     }
 }
 
 - (void)scrollViewDidEndDecelerating:(UIScrollView *)scrollView
 {
-    SmilieKeyboardSetScrollFractionForSmilieList(self.selectedSmilieList, scrollView.contentOffset.y / scrollView.contentSize.height);
+    SmilieKeyboardSetScrollFractionForSmilieList(self.selectedSmilieList, (float)scrollView.contentOffset.y / (float)scrollView.contentSize.height);
 }
 
 #pragma mark - UIInputViewAudioFeedback
